+++ src/containerapp/HISTORY.rst
@@ -5,8 +5,9 @@ Release History
 upcoming
 ++++++
 * 'az containerapp create': Fix Role assignment error when the default Azure Container Registry could not be found
-* Upgrade api-version to 2024-10-02-preview
+*  Upgrade api-version to 2024-10-02-preview
 * 'az containerapp create/update': `--yaml` support property pollingInterval and cooldownPeriod
+* 'az containerapp sessionpool create': Support managed identity when create session pool with --mi-system-assigned --mi-user-assigned
 
 1.0.0b4
 ++++++